# [distance][perf] Use fewer math.Sqrt calls.