### What's being changed:
Use mathematical property of square root $\sqrt{x} \times \sqrt{y}$ is equal to $\sqrt{x \times y}$. In addition to one fewer expensive operation, this change improves precision, since every float64 operation increases error.

Unlike in math, in practice due to limited float64 precision this change may result in overflows for vectors that previously would work, but given that the max float64 is 1.7976931348623157e+308, this should hopefully not be a problem.

Benchmark results:
before
```
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkCosineSimilarity$ github.com/weaviate/weaviate/usecases/vectorizer

goos: darwin
goarch: arm64
pkg: github.com/weaviate/weaviate/usecases/vectorizer
BenchmarkCosineSimilarity-8     78623644            15.27 ns/op        0 B/op          0 allocs/op
PASS
ok      github.com/weaviate/weaviate/usecases/vectorizer    1.359s
```
after:
```
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkCosineSimilarity$ github.com/weaviate/weaviate/usecases/vectorizer

goos: darwin
goarch: arm64
pkg: github.com/weaviate/weaviate/usecases/vectorizer
BenchmarkCosineSimilarity-8     67631196            15.13 ns/op        0 B/op          0 allocs/op
PASS
ok      github.com/weaviate/weaviate/usecases/vectorizer    1.185s
```

It's just a marginal perf improvement, but since it's also good for precision, this may be a good change.